requirements.txt
# Requirements for the FastAPI application
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
SQLAlchemy==2.0.35
requests==2.32.3
pytz==2024.1
apscheduler==3.10.4